added tests for py functions
