+++ R/add_tables_alt_text.R
@@ -26,8 +26,18 @@ add_tables_alt_text <- function(
 
   log4r::info(.le$logger, paste0("Output document path set: ", docx_out))
 
-  script <- system.file("scripts/add_table_alt_text.py", package = "reportifyr")
-  args <- c("run", script, "-i", docx_in, "-o", docx_out, "-l", .le$log_file)
+  args <- c(
+    "run",
+    "-m",
+    "reportipyr.cli",
+    "add-table-alt-text",
+    "-i",
+    docx_in,
+    "-o",
+    docx_out,
+    "-l",
+    .le$log_file
+  )
 
   paths <- get_venv_uv_paths()
