+++ R/keep_caption_next.R
@@ -10,11 +10,16 @@ keep_caption_next <- function(docx_in, docx_out) {
 
   paths <- get_venv_uv_paths()
 
-  script <- system.file(
-    "scripts/keep_caption_next.py",
-    package = "reportifyr"
+  args <- c(
+    "run",
+    "-m",
+    "reportipyr.cli",
+    "keep-caption-next",
+    "-i",
+    docx_in,
+    "-o",
+    docx_out
   )
-  args <- c("run", script, "-i", docx_in, "-o", docx_out)
 
   log4r::debug(.le$logger, "Running keep caption next script")
   run_python_script(
