+++ inst/extdata/uv_setup.sh
