+++ inst/python/reportipyr/__init__.py
@@ -1,0 +1,3 @@
+"""reportipyr python package for reportifyr."""
+
+__all__ = []
