+++ inst/python/reportipyr/figures.py
@@ -1,8 +1,5 @@
 import os
-
-import helper
 import tempfile
-import argparse
 from typing import Optional
 
 from docx import Document
@@ -11,8 +8,11 @@ from docx.enum.text import WD_ALIGN_PARAGRAPH
 
 from PIL import Image, ImageDraw, ImageFont
 
-from parse_magic_string import parse_magic_string
-from rpfy_logger import setup_logger
+from .config import load_yaml
+from .magic import parse_magic_string
+from .logging import setup_logger
+from .util import check_duplicates, create_label
+
 
 def add_figure(
     docx_in: str,
@@ -33,11 +33,13 @@ def add_figure(
     # load config.yaml or set empty dict for defaults.
     if config_yaml is not None:
         logger.debug(f"Loading config from: {config_yaml}")
-        config = helper.load_yaml(config_yaml)
+        config = load_yaml(config_yaml)
     else:
         config = {}
 
-    magic_pattern = helper.get_magic_pattern()
+    from .magic import get_magic_pattern
+
+    magic_pattern = get_magic_pattern()
 
     found_magic_strings = []
 
@@ -48,9 +50,7 @@ def add_figure(
 
         matches = magic_pattern.findall(par.text)
         if matches:
-            helper.check_duplicates(
-                matches, f"figure names in paragraph {actual_index+1}", logger
-            )
+            check_duplicates(matches, f"figure names in paragraph {actual_index+1}", logger)
 
             for match in matches:
                 logger.debug(f"Processing magic string: {match}")
@@ -93,13 +93,6 @@ def add_figure(
                         # Insert new paragraph after the current paragraph
                         new_par = document.add_paragraph()
                         run = new_par.add_run()
-                        # can only use embedded_size if the args are there
-                        # args = {
-                        #    'file.ext': {'width': '5', 'height': '8'},
-                        #    'file2.ext': {'height': '8'},
-                        #    'file3.ext': {'width': '5'},
-                        #    'file4.ext': {}
-                        # }
 
                         # Move paragraph to correct position (after current paragraph)
                         parent = par._element.getparent()
@@ -141,7 +134,7 @@ def add_figure(
                             if set(figure_args[figure].keys()).intersection(
                                 ["width", "height"]
                             ):
-                                logger.debug(f"Using embedded size from magic string")
+                                logger.debug("Using embedded size from magic string")
                                 embedded_width = figure_args[figure].get("width")
                                 embedded_height = figure_args[figure].get("height")
                                 run.add_picture(
@@ -217,7 +210,7 @@ def add_label_to_image(image_path: str, index: int, logger) -> str:
     This function saves the updated image to tmp and returns
     the path to the temp image.
     """
-    label = helper.create_label(index)
+    label = create_label(index)
 
     # load in image and create draw object
     # and set font
@@ -265,29 +258,115 @@ def add_label_to_image(image_path: str, index: int, logger) -> str:
     return temp_path
 
 
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description="Add figures to input docx document")
-    parser.add_argument(
-        "-i", "--input", type=str, required=True, help="input docx file path"
-    )
-    parser.add_argument("-o", "--output", type=str, required=True, help="output docx")
-    parser.add_argument(
-        "-d", "--figure_dir", type=str, required=True, help="Path to figures directory"
-    )
-    parser.add_argument(
-        "-c", "--config", type=str, default=None, help="Config yaml path"
-    )
-    parser.add_argument("-w", "--width", type=float, default=None, help="Figure width")
-    parser.add_argument("-g", "--height", type=float, default=None, help="Figure height")
-    parser.add_argument("-l", "--log", type=str, default=None, help="Log file path")
-    args = parser.parse_args()
+def remove_figures(
+    docx_in: str,
+    docx_out: str,
+    config_yaml: Optional[str],
+):
+    doc = Document(docx_in)
+    paragraphs = doc.paragraphs
 
-    add_figure(
-        args.input,
-        args.output,
-        args.figure_dir,
-        args.config,
-        args.width,
-        args.height,
-        args.log,
+    # load config.yaml or set empty dict for defaults.
+    if config_yaml is not None:
+        config = load_yaml(config_yaml)
+    else:
+        config = {}
+
+    for i, paragraph in enumerate(paragraphs):
+        text = paragraph.text.strip()
+        if text.startswith("{rpfy}:"):
+            figure_args = parse_magic_string(text)
+            update_magic_string = False
+
+            paragraphs_to_remove = []
+            for j, args in enumerate(figure_args.values()):
+                if i + j + 1 < len(paragraphs):
+                    next_par = paragraphs[i + j + 1]
+                    if not next_par.text.strip() and next_par._element.xpath(
+                        ".//w:drawing"
+                    ):
+                        paragraphs_to_remove.append((i + j + 1, next_par))
+                        if config.get("use_embedded_dimensions", True):
+                            dimensions = get_figure_dimensions(next_par)
+                            # set width and height from emu to Inches
+                            if dimensions.get("width"):
+                                args["width"] = str(
+                                    round(dimensions["width"] / 914400, 2)
+                                )
+                                update_magic_string = True
+                            if dimensions.get("height"):
+                                args["height"] = str(
+                                    round(dimensions["height"] / 914400, 2)
+                                )
+                                update_magic_string = True
+
+            if update_magic_string:
+                new_magic_string = "{rpfy}:"
+                if len(figure_args) > 1:
+                    new_magic_string += "["
+                    ending_string = "]"
+                else:
+                    ending_string = ""
+                for fig_idx, (fig, arg) in enumerate(figure_args.items()):
+                    arg_string = "<"
+                    for p_idx, (prop, val) in enumerate(arg.items()):
+                        arg_string += f"{prop}: {val}"
+                        if p_idx + 1 != len(arg):
+                            arg_string += ", "
+                    arg_string += ">"
+
+                    new_magic_string += f"{fig}{arg_string}"
+                    if fig_idx + 1 != len(figure_args):
+                        new_magic_string += ", "
+
+                new_magic_string += ending_string
+                paragraph.text = new_magic_string
+
+            for _, par in reversed(paragraphs_to_remove):
+                par._element.getparent().remove(par._element)
+
+    doc.save(docx_out)
+
+
+def get_figure_dimensions(paragraph) -> dict[str, Optional[int]]:
+    """Extract width and height from a paragraph containing a drawing."""
+    # Get all drawing elements in the paragraph
+    drawing_elements = paragraph._element.xpath(".//w:drawing")
+
+    if not drawing_elements:
+        return {}
+
+    # Get the first drawing element
+    drawing = drawing_elements[0]
+
+    # Find the extent element within the drawing
+    extent_elements = drawing.xpath(".//wp:extent")
+
+    if not extent_elements:
+        return {}
+
+    # Get the first extent element
+    extent = extent_elements[0]
+
+    # Extract cx and cy values
+    cx = extent.get(
+        "{http://schemas.openxmlformats.org/drawingml/2006/wordprocessingDrawing}cx"
     )
+    cy = extent.get(
+        "{http://schemas.openxmlformats.org/drawingml/2006/wordprocessingDrawing}cy"
+    )
+
+    # If using namespaces doesn't work, try getting attributes directly
+    if cx is None:
+        cx = extent.get("cx")
+    if cy is None:
+        cy = extent.get("cy")
+
+    if cx is None or cy is None:
+        return {}
+
+    # Convert to integers
+    width = int(cx)
+    height = int(cy)
+
+    return {"width": width, "height": height}
