+++ inst/python/tests/conftest.py
@@ -1,0 +1,8 @@
+import os
+import sys
+from pathlib import Path
+
+# Ensure reportipyr package is importable without installation.
+ROOT = Path(__file__).resolve().parents[1]
+if str(ROOT) not in sys.path:
+    sys.path.insert(0, str(ROOT))
